+++ bin/lib/App.js
@@ -1,14 +1,16 @@
 const util = require('util');
 const events = require('events');
 
-var App = function(logger) {
+var App = function(logger, logFile) {
 
   process.title = 'Lox2MQTT'; // required to restart process at OS level
 
   var that = this;
   this.logger = logger;
+  this.logFile = logFile;
 
   this.logger.info('Lox2MQTT started');
+  this.logger.startLog('lox2mqtt', 'Lox2MQTT', this.logFile, 'Lox2MQTT started');
 
   process.on('SIGINT', function() {
     that.logger.info('Lox2MQTT try to stop');
@@ -29,7 +31,7 @@ App.prototype.exit = function(code, message) {
   this.emit('exit', code);
 
   that.logger.info('Lox2MQTT exit - ' + message);
-  that.logger.closeLog();
+  that.logger.closeLog(that.logFile);
 };
 
 module.exports = App;